# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TripPlanner Kernel API",
    "description": "Kernel MVP API over local curated sample data.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/plan": {
      "post": {
        "summary": "Plan",
        "operationId": "plan_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripIntakeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Plan Plan Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plan/{job_id}": {
      "get": {
        "summary": "Get Job Status",
        "operationId": "get_job_status_plan__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanJobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plan/recompute": {
      "post": {
        "summary": "Recompute Plan",
        "operationId": "recompute_plan_plan_recompute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecomputeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plan/refresh-prose": {
      "post": {
        "summary": "Refresh Prose Plan",
        "operationId": "refresh_prose_plan_plan_refresh_prose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshProseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DraftItinerary-Input": {
        "properties": {
          "hotel_area_id": {
            "type": "string",
            "title": "Hotel Area Id"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/ItineraryDay-Input"
            },
            "type": "array",
            "minItems": 1,
            "title": "Days"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "itinerary_quality": {
            "type": "string",
            "enum": [
              "llm",
              "fallback"
            ],
            "title": "Itinerary Quality",
            "default": "llm"
          },
          "unverified_suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unverified Suggestions"
          }
        },
        "type": "object",
        "required": [
          "hotel_area_id",
          "days"
        ],
        "title": "DraftItinerary"
      },
...
      "FinalReport": {
        "properties": {
          "trip_spec": {
            "$ref": "#/components/schemas/TripSpec"
          },
          "itinerary": {
            "$ref": "#/components/schemas/DraftItinerary-Output"
          },
          "hotel_area": {
            "$ref": "#/components/schemas/SelectedHotelArea"
          },
          "flights_pick": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SampleFlight"
              },
              {
                "type": "null"
              }
            ]
          },
          "hotel_pick": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SampleHotel"
              },
              {
                "type": "null"
              }
            ]
          },
          "costed_trip": {
            "$ref": "#/components/schemas/CostedTrip"
          },
          "optimizer_result": {
            "$ref": "#/components/schemas/OptimizerResult"
          },
          "budget_totals": {
            "$ref": "#/components/schemas/BudgetTotals"
          },
          "payment_strategy": {
            "items": {
              "$ref": "#/components/schemas/PaymentStrategyRow"
            },
            "type": "array",
            "title": "Payment Strategy"
          },
          "transfer_advice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransferAdvice"
              },
              {
                "type": "null"
              }
            ]
          },
          "booking_checklist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Booking Checklist"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "provenance_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Provenance Warnings"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "caveats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Caveats"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "itinerary_overview": {
            "type": "string",
            "title": "Itinerary Overview",
            "default": ""
          },
          "payment_overview": {
            "type": "string",
            "title": "Payment Overview",
            "default": ""
          },
          "footer": {
            "type": "string",
            "title": "Footer",
            "default": ""
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "status": {
            "$ref": "#/components/schemas/PipelineStatus",
            "default": "ok"
          },
          "region_capability": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegionCapability"
              },
              {
                "type": "null"
              }
            ]
          },
          "freshness": {
            "$ref": "#/components/schemas/SectionFreshness"
          }
        },
        "type": "object",
        "required": [
          "trip_spec",
          "itinerary",
          "hotel_area",
          "costed_trip",
          "optimizer_result",
          "budget_totals",
          "payment_strategy",
          "confidence",
          "trace_id"
        ],
        "title": "FinalReport"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "JobError"
      },
...
      "PlanJobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "needs_clarification",
              "complete",
              "failed"
            ],
            "title": "Status"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "intake",
                  "itinerary",
                  "costing",
                  "optimizing",
                  "transfer",
                  "critic",
                  "explaining"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "stage_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Index"
          },
          "stages_total": {
            "type": "integer",
            "title": "Stages Total",
            "default": 6
          },
          "unresolved": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unresolved"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinalReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "PlanJobStatus"
      },
...
      "RecomputeRequest": {
        "properties": {
          "trip_spec": {
            "$ref": "#/components/schemas/TripSpec"
          },
          "itinerary": {
            "$ref": "#/components/schemas/DraftItinerary-Input"
          },
          "edit": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MoveItem"
              },
              {
                "$ref": "#/components/schemas/RemoveItem"
              },
              {
                "$ref": "#/components/schemas/ReorderDay"
              },
              {
                "$ref": "#/components/schemas/AddItem"
              },
              {
                "$ref": "#/components/schemas/ReplaceItem"
              }
            ],
            "title": "Edit",
            "discriminator": {
              "propertyName": "op",
              "mapping": {
                "add_item": "#/components/schemas/AddItem",
                "move_item": "#/components/schemas/MoveItem",
                "remove_item": "#/components/schemas/RemoveItem",
                "reorder_day": "#/components/schemas/ReorderDay",
                "replace_item": "#/components/schemas/ReplaceItem"
              }
            }
          },
          "previous_freshness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SectionFreshness"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "trip_spec",
          "itinerary",
          "edit"
        ],
        "title": "RecomputeRequest"
      },
...
      "RefreshProseRequest": {
        "properties": {
          "trip_spec": {
            "$ref": "#/components/schemas/TripSpec"
          },
          "itinerary": {
            "$ref": "#/components/schemas/DraftItinerary-Input"
          },
          "kernel_result": {
            "$ref": "#/components/schemas/KernelResult"
          },
          "previous_freshness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SectionFreshness"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "trip_spec",
          "itinerary",
          "kernel_result"
        ],
        "title": "RefreshProseRequest"
      },
...
      "TripIntakeRequest": {
        "properties": {
          "raw_request": {
            "type": "string",
            "title": "Raw Request"
          },
          "wallet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserWallet"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "raw_request"
        ],
        "title": "TripIntakeRequest"
      },
      "TripSpec": {
        "properties": {
          "home_country": {
            "type": "string",
            "enum": [
              "IN",
              "AE",
              "US"
            ],
            "title": "Home Country"
          },
          "origin_city": {
            "type": "string",
            "title": "Origin City"
          },
          "destination_city": {
            "type": "string",
            "title": "Destination City"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "travelers": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Travelers"
          },
          "budget_minor": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Minor"
          },
          "budget_currency": {
            "type": "string",
            "title": "Budget Currency",
            "default": "INR"
          },
          "style": {
            "type": "string",
            "enum": [
              "budget",
              "balanced",
              "luxury"
            ],
            "title": "Style"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Interests"
          },
          "pace": {
            "type": "string",
            "enum": [
              "relaxed",
              "moderate",
              "packed"
            ],
            "title": "Pace",
            "default": "moderate"
          },
          "dietary": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dietary"
          },
          "wallet": {
            "$ref": "#/components/schemas/UserWallet"
          },
          "optimization": {
            "$ref": "#/components/schemas/OptimizationPrefs"
          },
          "unresolved": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unresolved"
          }
        },
        "type": "object",
        "required": [
          "home_country",
          "origin_city",
          "destination_city",
          "start_date",
          "end_date",
          "travelers",
          "style",
          "wallet"
        ],
        "title": "TripSpec"
      },
...
      "UserWallet": {
        "properties": {
          "card_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Card Ids"
          },
          "points_balances": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Points Balances"
          }
        },
        "type": "object",
        "required": [
          "card_ids"
        ],
        "title": "UserWallet",
        "description": "The user's owned cards + optional points balances (from TripSpec, spec 01 \u00a78)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}